sucure the application

+++ app/api/issues/[id]/route.ts
@@ -1,51 +1,57 @@
+import authOptions from "@/app/auth/authOptions";
 import { issueSchema } from "@/app/validationSchema";
 import prisma from "@/prisma/client";
-import delay from "delay";
+import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function PATCH(
-    request : NextRequest,
-    { params }: { params: { id: string}}
-){
-    const body = await request.json()
-    const validation = issueSchema.safeParse(body)
+  request: NextRequest,
+  { params }: { params: { id: string } }
+) {
+  const session = await getServerSession(authOptions);
 
-    if(!validation.success) 
-        return NextResponse.json(validation.error.format(), { status:400});
-    
-    const issue = await prisma.issue.findUnique({
-        where:{ id: parseInt(params.id)}
-    })
+  if (!session) return NextResponse.json({}, { status: 401 });
+  const body = await request.json();
+  const validation = issueSchema.safeParse(body);
 
-    if(!issue)
-        return NextResponse.json({error: 'Invalid issue'} , { status: 404});
-    
-    const updateIssue = await prisma.issue.update({
-        where: { id : issue.id},
-        data: {
-            title: body.title,
-            description: body.description
-        }
-    })
+  if (!validation.success)
+    return NextResponse.json(validation.error.format(), { status: 400 });
 
-    return NextResponse.json(updateIssue)
+  const issue = await prisma.issue.findUnique({
+    where: { id: parseInt(params.id) },
+  });
+
+  if (!issue)
+    return NextResponse.json({ error: "Invalid issue" }, { status: 404 });
+
+  const updateIssue = await prisma.issue.update({
+    where: { id: issue.id },
+    data: {
+      title: body.title,
+      description: body.description,
+    },
+  });
+
+  return NextResponse.json(updateIssue);
 }
 
 export async function DELETE(
-    request : NextRequest,
-    { params }: { params: { id: string}}
-){
-    
-    const issue = await prisma.issue.findUnique({
-        where: { id: parseInt(params.id)}
-    })
+  request: NextRequest,
+  { params }: { params: { id: string } }
+) {
+  const session = await getServerSession(authOptions);
 
-    if(!issue)
-        return NextResponse.json({error: 'Invalid issue'}, { status:404})
+  if (!session) return NextResponse.json({}, { status: 401 });
+  const issue = await prisma.issue.findUnique({
+    where: { id: parseInt(params.id) },
+  });
 
-    await prisma.issue.delete({
-            where:{id: issue.id}
-    })
+  if (!issue)
+    return NextResponse.json({ error: "Invalid issue" }, { status: 404 });
 
-    return NextResponse.json({})
+  await prisma.issue.delete({
+    where: { id: issue.id },
+  });
+
+  return NextResponse.json({});
 }
